switch to MySQL backend
MongoDB has a hard limit on entries at 16MiB, which prevents clients
with the 45MiB libbulkextractor from being stored.  It is assumed that
the MySQL root password is 'root'.  MySQL must be set up manually before
initializing the server, and the packet size limit must be raised above
16M.

Additionally an apparent bug in the MySQL backend where null timestamp
values break insertion is hackily fixed by using the current time if the
timestamp is None